Add `ejira-heading-to-story-with-subtasks` function
As the name suggests, this function creates a story with subtasks.
Subtasks are created from the subheadings of the given header. This
function implemented using the primitives found in the package, I did
not do any api-level stuff, so it's kinda slow in a sense that firstly
it sends a request to create the story, then it sends a different
request for creating each subtask. I'm not sure if this can be reduced
into one or two requests but I'm happy with the given implementation.