provide overrides for JQL to sync tickets
This provides the ability to override the JQL used in ejira-update-jql.
From #20 there is a need to synchronize only the user's involved
tickets.

The original behavior should be unchanged. @osktyn has generously
provided a JQL statement to use. This is provided as
`ejira-jql-my-unresolved-fn`, which can be set via the variable
`ejira-update-jql-unresolved-fn`. An additional
`ejira-update-jql-resolved-fn` is also provided but doesn't need to be
overridden for the purposes of #20 - it just seemed appropriate given
this change.

Any stylistic changes are unintentional. Please nitpick away :)

Thanks for making an awesome project!

Closes #20